requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
filelock==3.3.0
huggingface-hub==0.0.19
idna==3.2
joblib==1.1.0
numpy==1.21.2
packaging==21.0
pyparsing==2.4.7
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
sacremoses==0.0.46
six==1.16.0
tokenizers==0.10.3
torch==1.9.1
tqdm==4.62.3
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
